requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.9
django-bootstrap3==9.1.0
django-toolbelt==0.0.1
gunicorn==19.7.1
psycopg2==2.7.3.2
static3==0.7.0
whitenoise==3.3.1
django-tinymce==2.6.0
